theme_news_agent/pyproject.toml
[tool.poetry]
name = "theme-news-agent"
version = "0.1.0"
description = "LLM Trend Theme Extraction AI Agent"
authors = ["RichardSeanPark <your_email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google-adk = "^0.2.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
playwright = "^1.51.0"
newspaper3k = "^0.2.8"
numpy = "^2.2.5"
pandas = "^2.2.3"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
pytrends = "^4.9.2"
lxml = {extras = ["html-clean"], version = "^5.3.2"}

[tool.poetry.group.dev.dependencies]
toml = "^0.10.2"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 